release.sh and GitHub workflow for releases
This adds a script to tag a release and keep version information in
sync, and a GitHub workflow to publish releases to crates.io and upload
binaries to a GitHub release if the crate is a `bin` crate.

This appends .liquid to the names of the GitHub workflow definition
files so that they don’t run on the template repo.