Use cargo-lints to centrally configure lints
Previously lints had to be configured at the top of every entry point in
rust code, e.g. `lib.rs`, `main.rs`, and each test file. This allows
lints to be centrally configured in `lints.toml` and ensures they are
applied to all targets.

To run lints:

    cargo lints clippy --all-targets --all-features

This requires [cargo-lints] to be installed.

[cargo-lints]: https://crates.io/crates/cargo-lints